requirements.txt
pydantic
certifi
aiohttp
pytest
pytest-asyncio
pytest-xdist
aleph-sdk-python==0.6.0